requirements.txt
mne>=1.7.0
mne-bids>=0.16.0
numpy>=1.26
scipy>=1.12
pandas>=2.2
scikit-learn>=1.5
xgboost>=2.1
pyyaml>=6.0
statsmodels>=0.14
matplotlib>=3.9

plotly>=5.24
